// src/module/fees/fine/dto/fine-collected.dto.ts
export class FineCollectedDto {
    id?: number;
  
    studentid: number;
  
    classId: number;
  
    feeMasterId: number;
  
    fineAmount: number;
  
    amountPaid: number;
  
    deductionAllowed: number;
  
    installment: number;
  
    createdDate?: Date;
  
    fromDate: Date;
  
    todate: Date;
  
    voucherEntryId: number;
}